+++ src/test/java/ru/mystamps/web/test/integration/cucumber/step/StepDefinitions.java
@@ -15,17 +15,18 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package ru.mystamps.web.tests;
+package ru.mystamps.web.test.integration.cucumber.step;
 
 import cucumber.api.java.en.And;
 import cucumber.api.java.en.Given;
 import cucumber.api.java.en.Then;
 import cucumber.api.java.en.When;
+
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.htmlunit.HtmlUnitDriver;
-import ru.mystamps.web.Url;
 
+import ru.mystamps.web.Url;
 import static org.junit.Assert.assertEquals;
 
 /**
